mk - calculator frontend, react-django integration
- calculator functionality works
- can switch between displaying either the calculator practice
  application or the author/quote tutorial applicaion
- demonstrates procedure to post data to backend and retrieve a response using axios and react
