+++ backend/core/models.py
@@ -1,3 +1,5 @@
+# models.py
+
 from django.db import models
 
 # Create your models here.
