+++ frontend/src/App.css
@@ -1,1 +1,7 @@
 /* App.css */
+
+.mainContainer {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+}
